@@ -5,7 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
 /**
@@ -40,19 +40,19 @@ public class Comment {
     /**
      * 用户ID
      */
-    @Column(nullable = false)
+    @Column(nullable = false,name = "user_id")
     private Long userId;
 
     /**
      * 作品ID（可为空）
      */
-    @Column
+    @Column(name = "artwork_id")
     private Long artworkId;
 
     /**
      * 话题ID（可为空）
      */
-    @Column
+    @Column(name = "topic_id")
     private Long topicId;
 
     /**
@@ -64,13 +64,13 @@ public class Comment {
     /**
      * 创建时间
      */
-    @Column(nullable = false, updatable = false)
+    @Column(nullable = false, updatable = false,name = "created_at")
     private LocalDateTime createdAt;
 
     /**
      * 更新时间
      */
-    @Column(nullable = false)
+    @Column(nullable = false,name = "updated_at")
     private LocalDateTime updatedAt;
 
     /**
