@@ -2,6 +2,7 @@ package com.zhuang.embroidery.aspect;
 
 import com.zhuang.embroidery.annotation.OperationLog;
 import com.zhuang.embroidery.service.OperationLogService;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.JoinPoint;
@@ -11,7 +12,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
 
-import javax.servlet.http.HttpServletRequest;
+
 import java.lang.reflect.Method;
 import java.util.HashMap;
 import java.util.Map;
